+++ src/app/views/demo/demo-routing.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HongkongPdComponent } from './map/hongkong-pd/hongkong-pd.component';
-import { BusLineComponent } from './map/bus-line/bus-line.component';
 import { OfficialDemoComponent } from './tree/official-demo/official-demo.component';
 
 const routes: Routes = [
@@ -10,13 +9,6 @@ const routes: Routes = [
     component: HongkongPdComponent,
     data: {
       title: 'Population Density of HongKong (2011)'
-    }
-  },
-  {
-    path: 'map/bus-lines-of-beijing',
-    component: BusLineComponent,
-    data: {
-      title: 'Bus Lines of Beijing (Baidu Map)'
     }
   },
   {