+++ src/app/views/usage/usage.module.ts
@@ -9,7 +9,6 @@ import { InitOptsComponent } from './init-opts/init-opts.component';
 import { LoadingComponent } from './loading/loading.component';
 import { EchartsInstanceComponent } from './echarts-instance/echarts-instance.component';
 import { EventsComponent } from './events/events.component';
-import { NgxEchartsServiceComponent } from './ngx-echarts-service/ngx-echarts-service.component';
 import { AngularDraggableModule } from 'angular2-draggable';
 import { AutoResizeComponent } from './auto-resize/auto-resize.component';
 
@@ -28,7 +27,6 @@ import { AutoResizeComponent } from './auto-resize/auto-resize.component';
     LoadingComponent,
     EchartsInstanceComponent,
     EventsComponent,
-    NgxEchartsServiceComponent,
     AutoResizeComponent
   ]
 })