+++ src/main.ts
@@ -4,6 +4,11 @@ import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
 
+/** echarts extensions: */
+import 'echarts/theme/dark.js';
+import 'echarts/theme/macarons.js';
+import 'echarts/dist/extension/bmap.min.js';
+
 if (environment.production) {
   enableProdMode();
 }